Move checks and transforms to SDK
For now the old modules remain and raise deprecation warnings, we can
remove these later when the repositories are up to date.

Fixes #1117
