+++ backend/tests/fixtures/project_02/checks/check_spine_interface_status.py
@@ -1,4 +1,4 @@
-from infrahub.checks import InfrahubCheck
+from infrahub_sdk.checks import InfrahubCheck
 
 
 class InfrahubCheckSpineNbrInterfaceDisabled(InfrahubCheck):
