+++ backend/tests/fixtures/transforms/transform01.py
@@ -1,4 +1,4 @@
-from infrahub.transforms import InfrahubTransform
+from infrahub_sdk.transforms import InfrahubTransform
 
 
 class Transform01(InfrahubTransform):
