+++ backend/tests/fixtures/transforms/transform02.py
@@ -1,9 +1,8 @@
-from infrahub.transforms import InfrahubTransform
+from infrahub_sdk.transforms import InfrahubTransform
 
 
 class Transform02(InfrahubTransform):
     query = "my_query"
-    # url = "transform01"
 
     def transform(self, data: dict):
         return {str(key).upper(): value for key, value in data.items()}
