+++ python_sdk/tests/unit/sdk/checks/test_checks.py
@@ -1,7 +1,7 @@
 import pytest
-from infrahub_sdk import InfrahubClient
 
-from infrahub.checks import InfrahubCheck
+from infrahub_sdk import InfrahubClient
+from infrahub_sdk.checks import InfrahubCheck
 
 
 async def test_class_init():
